+ In order to support Fund purchases which don't have to be integer amounts (unlike individual stocks or ETF's) the inputs now support decimals up to six decimal places for increased precision.

+++ app/src/main/java/uk/co/nyakeh/stacks/database/StockPurchaseCursorWrapper.java
@@ -19,7 +19,7 @@ public class StockPurchaseCursorWrapper extends CursorWrapper {
         String symbol = getString(getColumnIndex(StockPurchaseTable.Cols.SYMBOL));
         Long datePurchased = getLong(getColumnIndex(StockPurchaseTable.Cols.DATE_PURCHASED));
         double price = getDouble(getColumnIndex(StockPurchaseTable.Cols.PRICE));
-        int quantity = getInt(getColumnIndex(StockPurchaseTable.Cols.QUANTITY));
+        double quantity = getDouble(getColumnIndex(StockPurchaseTable.Cols.QUANTITY));
         double fee = getDouble(getColumnIndex(StockPurchaseTable.Cols.FEE));
         double total = getDouble(getColumnIndex(StockPurchaseTable.Cols.TOTAL));
